# backend/src/swaggerSpec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cultural Sites API",
    "version": "1.0.0",
    "description": "API documentation for Cultural Sites"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/reviews/site/{siteId}": {
      "get": {
        "tags": ["Review"],
        "summary": "Get reviews for a specific site",
        "parameters": [
          {
            "name": "siteId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID of the site to fetch reviews for"
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews for the site"
          },
          "500": {
            "description": "Server error while fetching reviews"
          }
        }
      },
      "post": {
        "tags": ["Review"],
        "summary": "Create or update a review for a specific site",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "siteId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID of the site to review"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
                  "comment": { "type": "string" }
                },
                "required": ["rating"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created"
          },
          "200": {
            "description": "Review updated"
          },
          "400": {
            "description": "Validation error (e.g., rating out of range)"
          },
          "401": {
            "description": "User not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/reviews/{reviewId}": {
      "put": {
        "tags": ["Review"],
        "summary": "Update an existing review",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID of the review to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
                  "comment": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Review not found or unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": ["Review"],
        "summary": "Delete a review",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID of the review to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Review not found or unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/reviews/user": {
      "get": {
        "tags": ["Review"],
        "summary": "Get reviews by the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of user reviews"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/favorites": {
      "get": {
        "tags": ["Favorites"],
        "summary": "Get all favorites for the current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of favorite sites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Favorite"
                  }
                }
              }
            }
          },
          "401": { "description": "Not authenticated" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/favorites/{siteId}": {
      "post": {
        "tags": ["Favorites"],
        "summary": "Add a site to favorites",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "siteId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "201": {
            "description": "Site added to favorites",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Favorite" }
              }
            }
          },
          "401": { "description": "Not authenticated" },
          "409": { "description": "Site already in favorites" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "tags": ["Favorites"],
        "summary": "Remove a site from favorites",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "siteId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Favorite removed successfully" },
          "401": { "description": "Not authenticated" },
          "404": { "description": "Favorite not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/favorites/{siteId}/check": {
      "get": {
        "tags": ["Favorites"],
        "summary": "Check if a site is favorited by the current user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "siteId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorite status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "is_favorited": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "description": "Not authenticated" },
          "500": { "description": "Server error" }
        }
      }
    },
...
  }
}